file sharing indirection to keep metadata synced

+++ shared/js/cyph/services/account-files.service.ts
@@ -9,7 +9,14 @@ import {DeltaOperation, DeltaStatic} from 'quill';
 import * as Delta from 'quill-delta';
 import * as QuillDeltaToHtml from 'quill-delta-to-html';
 import {BehaviorSubject, Observable} from 'rxjs';
-import {AccountFileRecord, Form, IAccountFileRecord, IForm} from '../../proto';
+import {
+	AccountFileRecord,
+	AccountFileReference,
+	Form,
+	IAccountFileRecord,
+	IAccountFileReference,
+	IForm
+} from '../../proto';
 import {SecurityModels} from '../account';
 import {IProto} from '../iproto';
 import {IQuillDelta} from '../iquill-delta';
@@ -37,17 +44,34 @@ export class AccountFilesService {
 	private readonly noteSnippets: Map<string, string>	= new Map<string, string>();
 
 	/** List of file records owned by current user, sorted by timestamp in descending order. */
-	public readonly filesList: Observable<IAccountFileRecord[]>	= util.flattenObservablePromise(
-		this.accountDatabaseService.watchList<IAccountFileRecord>(
-			'fileRecords',
-			AccountFileRecord
-		).map(records =>
-			records.
-				map(({value}) => value).
-				sort((a, b) => b.timestamp - a.timestamp)
-		),
-		[]
-	);
+	public readonly filesList: Observable<(IAccountFileRecord&{owner: string})[]>	=
+		util.flattenObservablePromise(
+			this.accountDatabaseService.watchList(
+				'fileReferences',
+				AccountFileReference
+			).flatMap(async references => Promise.all(references.map(async ({value}) => {
+				const record	= await this.accountDatabaseService.getItem(
+					`users/${value.owner}/fileRecords/${value.id}`,
+					AccountFileRecord,
+					undefined,
+					value.key
+				);
+
+				return {
+					id: record.id,
+					mediaType: record.mediaType,
+					name: record.name,
+					owner: value.owner,
+					recordType: record.recordType,
+					size: record.size,
+					timestamp: record.timestamp
+				};
+			}))).map(records =>
+				records.sort((a, b) => b.timestamp - a.timestamp)
+			),
+			[]
+		)
+	;
 
 	/**
 	 * Files filtered by record type.
@@ -102,11 +126,13 @@ export class AccountFilesService {
 		progress: Observable<number>;
 		result: Promise<T>;
 	} {
+		const filePromise	= this.getFile(id);
+
 		const {progress, result}	= this.accountDatabaseService.downloadItem(
-			(async () => `${(await this.getFile(id)).owner}/files/${id}`)(),
+			(async () => `${(await filePromise).owner}/files/${id}`)(),
 			proto,
 			securityModel,
-			(async () => (await this.getFile(id)).key)()
+			(async () => (await filePromise).key)()
 		);
 
 		return {progress, result: (async () => (await result).value)()};
@@ -131,10 +157,10 @@ export class AccountFilesService {
 	}
 
 	/** Accepts incoming file. */
-	public async acceptIncomingFile (incomingFile: IAccountFileRecord) : Promise<void> {
+	public async acceptIncomingFile (incomingFile: IAccountFileReference) : Promise<void> {
 		await this.accountDatabaseService.pushItem(
-			`fileRecords/${incomingFile.id}`,
-			AccountFileRecord,
+			`fileReferences/${incomingFile.id}`,
+			AccountFileReference,
 			incomingFile
 		);
 
@@ -182,17 +208,33 @@ export class AccountFilesService {
 	public async getFile (
 		id: string,
 		recordType?: AccountFileRecord.RecordTypes
-	) : Promise<IAccountFileRecord> {
-		const file	= await this.accountDatabaseService.getItem(
-			`fileRecords/${id}`,
-			AccountFileRecord
+	) : Promise<IAccountFileRecord&IAccountFileReference> {
+		const reference	= await this.accountDatabaseService.getItem(
+			`fileReferences/${id}`,
+			AccountFileReference
 		);
 
-		if (recordType !== undefined && file.recordType !== recordType) {
+		const record	= await this.accountDatabaseService.getItem(
+			`users/${reference.owner}/fileRecords/${reference.id}`,
+			AccountFileRecord,
+			undefined,
+			reference.key
+		);
+
+		if (recordType !== undefined && record.recordType !== recordType) {
 			throw new Error('Specified file does not exist.');
 		}
 
-		return file;
+		return {
+			id,
+			key: reference.key,
+			mediaType: record.mediaType,
+			name: record.name,
+			owner: reference.owner,
+			recordType: record.recordType,
+			size: record.size,
+			timestamp: record.timestamp
+		};
 	}
 
 	/** Gets the Material icon name for the file default thumbnail. */
@@ -218,13 +260,14 @@ export class AccountFilesService {
 
 			(async () => {
 				const limit		= 75;
+				const file		= await this.getFile(id);
 				const content	= this.deltaToString(
 					msgpack.decode(
 						await this.accountDatabaseService.getItem(
-							`files/${id}`,
+							`users/${file.owner}/files/${id}`,
 							BinaryProto,
 							undefined,
-							(await this.getFile(id)).key
+							file.key
 						)
 					)
 				);
@@ -265,9 +308,9 @@ export class AccountFilesService {
 			id	= (await id).id;
 		}
 
-		if (confirmAndRedirect) {
-			const file	= await this.getFile(id);
+		const file	= await this.getFile(id);
 
+		if (confirmAndRedirect) {
 			if (await this.dialogService.confirm({
 				content: `${this.stringsService.deleteMessage} ${file.name}?`,
 				title: this.stringsService.deleteConfirm
@@ -291,9 +334,10 @@ export class AccountFilesService {
 		}
 
 		await Promise.all([
-			this.accountDatabaseService.removeItem(`docs/${id}`),
-			this.accountDatabaseService.removeItem(`files/${id}`),
-			this.accountDatabaseService.removeItem(`fileRecords/${id}`)
+			this.accountDatabaseService.removeItem(`users/${file.owner}/docs/${id}`),
+			this.accountDatabaseService.removeItem(`users/${file.owner}/files/${id}`),
+			this.accountDatabaseService.removeItem(`users/${file.owner}/fileRecords/${id}`),
+			this.accountDatabaseService.removeItem(`fileReferences/${id}`)
 		]);
 	}
 
@@ -303,7 +347,7 @@ export class AccountFilesService {
 			`users/${username}/incomingFiles/${id}`,
 			BinaryProto,
 			await this.potassiumService.box.seal(
-				await this.accountDatabaseService.getItem(`fileRecords/${id}`, BinaryProto),
+				await this.accountDatabaseService.getItem(`fileReferences/${id}`, BinaryProto),
 				(await this.accountDatabaseService.getUserPublicKeys(username)).publicEncryptionKey
 			)
 		);
@@ -311,12 +355,13 @@ export class AccountFilesService {
 
 	/** Overwrites an existing note. */
 	public async updateDoc (id: string, delta: IQuillDelta|IQuillRange) : Promise<void> {
+		const file	= await this.getFile(id);
 		await this.accountDatabaseService.pushItem(
-			`docs/${id}`,
+			`users/${file.owner}/docs/${id}`,
 			BinaryProto,
 			msgpack.encode(delta),
 			undefined,
-			(await this.getFile(id)).key
+			file.key
 		);
 	}
 
@@ -327,14 +372,23 @@ export class AccountFilesService {
 		size?: number;
 	}) : Promise<void> {
 		const original	= await this.getFile(id);
-		await this.accountDatabaseService.setItem(`fileRecords/${id}`, AccountFileRecord, {
-			id,
-			mediaType: metadata.mediaType === undefined ? original.mediaType : metadata.mediaType,
-			name: metadata.name === undefined ? original.name : metadata.name,
-			recordType: original.recordType,
-			size: metadata.size === undefined ? original.size : metadata.size,
-			timestamp: await util.timestamp()
-		});
+		await this.accountDatabaseService.setItem(
+			`users/${original.owner}/fileRecords/${id}`,
+			AccountFileRecord,
+			{
+				id,
+				mediaType: metadata.mediaType === undefined ?
+					original.mediaType :
+					metadata.mediaType
+				,
+				name: metadata.name === undefined ? original.name : metadata.name,
+				recordType: original.recordType,
+				size: metadata.size === undefined ? original.size : metadata.size,
+				timestamp: await util.timestamp()
+			},
+			undefined,
+			original.key
+		);
 	}
 
 	/** Overwrites an existing note. */
@@ -349,13 +403,17 @@ export class AccountFilesService {
 
 		await Promise.all([
 			this.accountDatabaseService.setItem(
-				`files/${id}`,
+				`users/${file.owner}/files/${id}`,
 				BinaryProto,
 				msgpack.encode(content),
 				undefined,
-				(await this.getFile(id)).key
+				file.key
 			),
-			this.accountDatabaseService.setItem(`fileRecords/${id}`, AccountFileRecord, file)
+			this.accountDatabaseService.setItem(
+				`users/${file.owner}/fileRecords/${id}`,
+				AccountFileRecord,
+				file
+			)
 		]);
 	}
 
@@ -418,7 +476,6 @@ export class AccountFilesService {
 					AccountFileRecord,
 					{
 						id,
-						key: await key,
 						mediaType: file instanceof Blob ?
 							file.type :
 							file instanceof Array ?
@@ -428,7 +485,6 @@ export class AccountFilesService {
 									'cyph/form'
 						,
 						name,
-						owner: (await this.accountDatabaseService.getCurrentUser()).user.username,
 						recordType: file instanceof Blob ?
 							AccountFileRecord.RecordTypes.File :
 							file instanceof Array ?
@@ -446,6 +502,18 @@ export class AccountFilesService {
 								NaN
 						,
 						timestamp: await util.timestamp()
+					},
+					undefined,
+					key
+				);
+
+				await this.accountDatabaseService.setItem(
+					`fileReferences/${id}`,
+					AccountFileReference,
+					{
+						id,
+						key: await key,
+						owner: (await this.accountDatabaseService.getCurrentUser()).user.username
 					}
 				);
 
@@ -459,13 +527,19 @@ export class AccountFilesService {
 		deltas: Observable<IQuillDelta>;
 		selections: Observable<IQuillRange>;
 	}> {
-		const length	= (await this.accountDatabaseService.getListKeys(`docs/${id}`)).length;
+		const file		= await this.getFile(id);
+
+		const length	=
+			(
+				await this.accountDatabaseService.getListKeys(`users/${file.owner}/docs/${id}`)
+			).length
+		;
 
 		const doc		= this.accountDatabaseService.watchListPushes(
-			`docs/${id}`,
+			`users/${file.owner}/docs/${id}`,
 			BinaryProto,
 			undefined,
-			(await this.getFile(id)).key
+			file.key
 		).map(o =>
 			o.value.length > 0 ? msgpack.decode(o.value) : undefined
 		);
@@ -490,7 +564,7 @@ export class AccountFilesService {
 	/** Watches file record. */
 	public watchMetadata (id: string) : Observable<IAccountFileRecord> {
 		return this.accountDatabaseService.watch(
-			`fileRecords/${id}`,
+			(async () => `users/${(await this.getFile(id)).owner}/fileRecords/${id}`)(),
 			AccountFileRecord
 		).map(o =>
 			o.value
